# openapi.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "Product Controller"
        ],
        "summary": "Получить продукт",
        "description": "Возвращает информацию о продукте по его идентификатору.",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Продукт найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "404": {
            "description": "Продукт не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product Controller"
        ],
        "summary": "Обновить продукт",
        "description": "Обновляет данные продукта по его ID. Только для администраторов.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Продукт успешно обновлён",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "404": {
            "description": "Продукт не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product Controller"
        ],
        "summary": "Удалить продукт",
        "description": "Удаляет продукт по его ID. Только для администраторов.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Продукт успешно удалён"
          },
          "403": {
            "description": "Доступ запрещён"
          },
          "404": {
            "description": "Продукт не найден"
          }
        }
      }
    },
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "Product Controller"
        ],
        "summary": "Список продуктов",
        "description": "Возвращает список продуктов с возможностью фильтрации по категории и имени.",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список продуктов успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageableResponseProductDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product Controller"
        ],
        "summary": "Создать продукт",
        "description": "Создаёт новый продукт. Только для администраторов.",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Продукт успешно создан",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/change-price": {
      "post": {
        "tags": [
          "Product Price Controller"
        ],
        "summary": "Изменить цену",
        "description": "Изменяет цену на продукт. Только для администраторов.",
        "operationId": "changePrice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePriceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Цена успешно изменена"
          },
          "400": {
            "description": "Неверные данные"
          },
          "403": {
            "description": "Доступ запрещён"
          },
          "404": {
            "description": "Продукт или магазин не найдены"
          }
        }
      }
    },
...
    "/api/v1/products/{id}/stats": {
      "get": {
        "tags": [
          "Product Price Controller"
        ],
        "summary": "Статистика цен",
        "description": "Возвращает историю изменения цены на продукт за указанный период.",
        "operationId": "getStats",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Статистика успешно получена",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageableResponsePriceAtDate"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные параметры запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageableResponsePriceAtDate"
                }
              }
            }
          },
          "404": {
            "description": "Продукт не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageableResponsePriceAtDate"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/shop-prices": {
      "get": {
        "tags": [
          "Product Price Controller"
        ],
        "summary": "Список цен по магазинам",
        "description": "Возвращает текущие цены на продукт в разных магазинах.",
        "operationId": "listShopPrices",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Цены успешно получены",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageableResponseShopPriceDto"
                }
              }
            }
          },
          "404": {
            "description": "Продукт не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageableResponseShopPriceDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "info": {
            "type": "string",
            "minLength": 1
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "ChangePriceDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "shopId": {
            "type": "integer",
            "format": "int64"
          },
          "newPrice": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageableResponseProductDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "object"
          }
        }
      },
      "PageableResponsePriceAtDate": {
        "type": "object",
        "properties": {
          "items": {
            "type": "object"
          }
        }
      },
      "PageableResponseShopPriceDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}