# src/main/java/com/senla/pricemonitor/dto/ChangePriceDto.java
package com.senla.pricemonitor.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class ChangePriceDto {
    private Long productId;
    private Long shopId ;
    private Long newPrice;
}